MOD : Support for Abstract entity RestCollection
On supporte les collections de classe abstraite, et plus généralement les collections "read only"
Ca permet d'éclater en 2 les RestCollection, avec d'un côté ReadOnlyRestCollection qui contient tout ce qui est GET, et on ne laisse dans RestCollection que ce a trait aux PUT/POST/DELETE

Resolves #28
