+++ Domain/RDD.Domain.Mocks/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Moq" version="4.5.3" targetFramework="net461" />
+</packages>
