+++ RDD.Domain/RDD.Domain.csproj
@@ -95,6 +95,7 @@
     <Compile Include="IIncludable.cs" />
     <Compile Include="IPrimaryKey.cs" />
     <Compile Include="IPrincipal.cs" />
+    <Compile Include="IReadOnlyRestCollection.cs" />
     <Compile Include="IRestCollection.cs" />
     <Compile Include="ISelection.cs" />
     <Compile Include="IStorageService.cs" />
@@ -118,11 +119,12 @@
     <Compile Include="Models\Querying\SerilizationService.cs" />
     <Compile Include="Models\Querying\Where.cs" />
     <Compile Include="Exceptions\UnreachableEntityTypeException.cs" />
+    <Compile Include="Models\ReadOnlyRestCollection.cs" />
+    <Compile Include="Models\ReadOnlyRestCollection.queryable.cs" />
     <Compile Include="Models\Selection.cs" />
     <Compile Include="WebServices\WebService.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Models\RestCollection.cs" />
-    <Compile Include="Models\RestCollection.queryable.cs" />
     <Compile Include="WebServices\WebServicesCollection.cs" />
   </ItemGroup>
   <ItemGroup>
