+++ RDD.sln
@@ -13,6 +13,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RDD.Domain.Tests", "Domain\
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RDD.Infra.Tests", "RDD.Infra.Tests\RDD.Infra.Tests.csproj", "{25461E71-7E9B-47BA-8018-418F58BBBFAC}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{C7FDD302-B6F8-4E61-8995-112D2187B086}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RDD.Domain.Mocks", "Domain\RDD.Domain.Mocks\RDD.Domain.Mocks.csproj", "{B4991FA4-2578-45A2-9780-B4FA13F5E9CB}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -39,6 +43,10 @@ Global
 		{25461E71-7E9B-47BA-8018-418F58BBBFAC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{25461E71-7E9B-47BA-8018-418F58BBBFAC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{25461E71-7E9B-47BA-8018-418F58BBBFAC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B4991FA4-2578-45A2-9780-B4FA13F5E9CB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B4991FA4-2578-45A2-9780-B4FA13F5E9CB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B4991FA4-2578-45A2-9780-B4FA13F5E9CB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B4991FA4-2578-45A2-9780-B4FA13F5E9CB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
